REFACT: Ball 구조 일부 개선
